+++ legion/engine/llri-dx/detail/adapter.cpp
@@ -1,0 +1,66 @@
+#include <llri/llri.hpp>
+#include <graphics/directx/d3d12.h>
+#include <dxgi1_6.h>
+
+namespace legion::graphics::llri
+{
+    namespace internal
+    {
+        result mapHRESULT(const HRESULT& value);
+    }
+
+    result Adapter::impl_queryInfo(adapter_info* info) const
+    {
+        DXGI_ADAPTER_DESC1 desc;
+        static_cast<IDXGIAdapter1*>(m_ptr)->GetDesc1(&desc);
+
+        adapter_info result;
+        result.vendorId = desc.VendorId;
+        result.adapterId = desc.DeviceId;
+        const auto description = std::string(desc.Description, desc.Description + 128);
+        result.adapterName = description.substr(0, description.find_last_not_of(' '));
+
+        if (desc.Flags == DXGI_ADAPTER_FLAG_REMOTE)
+            result.adapterType = adapter_type::Virtual;
+        else if (desc.DedicatedVideoMemory > 0)
+            result.adapterType = adapter_type::Discrete;
+        else // video memory == 0 
+            result.adapterType = adapter_type::Integrated;
+
+        *info = result;
+        return result::Success;
+    }
+
+    result Adapter::impl_queryFeatures(adapter_features* features) const
+    {
+        HRESULT level12_0 = D3D12CreateDevice(static_cast<IDXGIAdapter*>(m_ptr), D3D_FEATURE_LEVEL_12_0, __uuidof(ID3D12Device), nullptr);
+        if (level12_0 == E_FAIL)
+            return result::ErrorIncompatibleDriver;
+        else if (FAILED(level12_0)) //no matter what reason, LEVEL_12_0 should always be supported for DX12
+            return internal::mapHRESULT(level12_0);
+
+        //Level 1 and 2 can more easily grant us guarantees about feature support
+        HRESULT level12_1 = D3D12CreateDevice(static_cast<IDXGIAdapter*>(m_ptr), D3D_FEATURE_LEVEL_12_0, __uuidof(ID3D12Device), nullptr);
+        HRESULT level12_2 = D3D12CreateDevice(static_cast<IDXGIAdapter*>(m_ptr), D3D_FEATURE_LEVEL_12_0, __uuidof(ID3D12Device), nullptr);
+
+        adapter_features output;
+
+        //Set all the information in a structured way here
+
+        *features = output;
+        return result::Success;
+    }
+
+    result Adapter::impl_queryExtensionSupport(const adapter_extension_type& type, bool* supported) const
+    {
+        *supported = false;
+
+        switch (type)
+        {
+        default:
+            break;
+        }
+
+        return result::Success;
+    }
+}
