+++ legion/engine/llri-dx/llri.cpp
@@ -1,0 +1,6 @@
+#include <llri/llri.hpp>
+
+namespace legion::graphics::llri
+{
+    //File reserved for future use
+}
